move some file to deprecated
